renamed convert.ts to uuidToURL.ts
